requirements.txt
openai
langchain
faiss-cpu
pdfplumber
fastapi
uvicorn
streamlit
tiktoken
python-dotenv